app/requirements.txt
fastapi
openai
pandas
uvicorn
pymupdf
python-docx
python-multipart
openpyxl
xlsxwriter